% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autota.R
\name{handle_error}
\alias{handle_error}
\title{Runs all error handlers against an rlang trace.
For internal use only.}
\usage{
handle_error(trace)
}
\arguments{
\item{trace}{the rlang trace to handle}
}
\description{
Runs all error handlers against an rlang trace.
For internal use only.
}
